Reduce noise in server RPC updates
Because RPC updates are tracked based on changes in the throttle,
the throttle could report (0, 0), (10, 0), (0, 0), each of these
being a valid update and setting the self.tx_rates for Control to
the proper amounts. However, this overall has the effect of being
a useless update to the client. As such we now track the last sent
tx_rate and make sure to deduplicate against that.