Remove Send impl for Throttle
While the Rc-wrapped data only appears to be used from one
thread, this Send impl makes it easy to accidentally cause UB
if Throttle is ever used before sending it across a channel.

Instead, only the TcpStream and Reader for a given peer are send across
a channel. The PeerConn and its internal Throttle are constructed on
the other side, allowing an Rc to be used without any potential for UB