Handle disk back pressure
Synapse now uses a fixed number of buffers for transferring data between
threads, stalling downloads if this buffer list becomes exhausted, which
may happen if the net speed is faster than the disk speed. This works
well for downloads, but needs refinement for uploads, as right now
requests which can't be fulfilled are dropped.

Closes #103