Implement torrent importing
Importing allows for adding torrents to be added without validation if
they already exist on the computer. Note that synapse will still check a
single piece of the torrent to ensure that it exists in some capacity,
though users should ensure that their torrents are valid before
regardless before using this functionality. This commit also includes a
reworking of the bencode parser so they now no longer recurse,
preventing potential overflows.