Fix metadata transfer
The previous changes to the BEncode decoder resulted in metadata
transfer breaking because its protocol relies on data appended to
a BEncoded message, which resulted in a bad decode.

Our BEncoding library now supports the ability to decode the first
valid BEncode message in a buffer allowing for these messages to be
parsed correctly.